Added links for meeting registration

+++ project/payment/forms.py
@@ -2,7 +2,7 @@
 
 
 from flask_wtf import Form
-from wtforms import StringField, FloatField, SelectField
+from wtforms import StringField, FloatField, SelectField, IntegerField
 from wtforms.validators import DataRequired, Length
 
 
@@ -11,7 +11,7 @@ class CardPaymentForm(Form):
     names = StringField('names', validators=[DataRequired(), Length(min=3, max=45)])
     cardNumber = StringField('cardNumber', validators=[DataRequired(), Length(min=6, max=25)], default="None")
     phone = StringField('phone', validators=[DataRequired(), Length(min=2, max=25)])
-    amount = FloatField('Amount', default=1.0, validators=[DataRequired("This field is required")])
+    amount = IntegerField('Amount', default=1, validators=[DataRequired("This field is required")])
     object_payment = SelectField('objectPayment', validators=[DataRequired("Please select the reason for payment")],
                                  choices=[('membership', 'Annual Membership'), ('donation', 'Donation'),
                                           ('meeting', 'Meeting Registration')])
